feat: Add AssignReviewersModel and AssignReviewers endpoint
The code changes introduce a new `AssignReviewersModel` class in the `DTOs` folder, which includes properties for `SubmissionId` and `ReviewerIds`. Additionally, the `ISubmissionService` interface is updated with a new method `AssignReviewersAsync` that takes an `int` submissionId and a list of reviewerIds.

This commit also includes the addition of an `AssignReviewers` endpoint in the `ManuscriptsController` that allows editors to assign reviewers to a submission.

```