refactor: Update ManuscriptsController to allow multiple roles for authorization
The code changes modify the `ManuscriptsController` to allow multiple roles for authorization. Previously, only the "Author" role was allowed to access the `UploadFile` endpoint. With this change, the roles "Author", "Reviewer", and "Editor" are now allowed to access the endpoint. This update enables a more flexible and inclusive access control for file uploads.