refactor: Add AuthorResponseDTO and RevisedFilePaths property
The code changes add a new class called `AuthorResponseDTO` in the `DTOs` folder. This class represents the response from an author for a submission. It includes properties for the submission ID and the response file.

In the `ManuscriptModel.cs` file, a new property `RevisedFilePaths` is added. This property is used to store the file paths of revised manuscripts.