refactor: Add RequestExtension endpoint to ManuscriptsController
The code changes in ManuscriptsController.cs add a new endpoint called RequestExtension. This endpoint allows users to request an extension for a submission by providing the submission ID. If the request is successful, it returns a success message; otherwise, it returns an error message. This update enhances the functionality of the manuscript submission process by allowing authors to request extensions when needed.